Add important information on InterfaceMapping.xml
The main thing of 'InterfaceMapping' is to find matching methods between an interface and a concrete class.
Mostly of times it is required to find the method pendant. So the Information that the order of the arrays is mirrored is heavy important (I wasted my time finding the pendant method).
At the moment the detail about the ordering is explained on the member's remarks text. But it should be on class summary too.